requirements.txt
# LangChain Core
langchain>=0.3.7
langchain-openai>=0.2.5
langchain-chroma>=0.1.4
langchain-community>=0.3.5

# Vector Stores
chromadb>=0.5.20
pinecone-client>=5.0.1

# OpenTelemetry for Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation>=0.48b0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-langchain>=0.32b0
opentelemetry-exporter-otlp>=1.27.0

# API Framework
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
pydantic>=2.10.2
pydantic-settings>=2.6.1

# Testing
pytest>=8.3.4
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
pytest-cov>=6.0.0

# Utilities
python-dotenv>=1.0.1
tiktoken>=0.8.0
httpx>=0.27.2

# Monitoring & Metrics
prometheus-client>=0.21.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=14.0.0

# HTTP & Data Ingestion
requests>=2.31.0
tqdm>=4.66.0
